Add TicketIssue component
