+++ client/src/app/pages/migration-waves/migration-waves.tsx
@@ -66,7 +66,6 @@ import { WaveForm } from "./components/migration-wave-form";
 import { ManageApplicationsForm } from "./components/manage-applications-form";
 import { deleteMigrationWave } from "@app/api/rest";
 import { ConditionalTooltip } from "@app/shared/components/ConditionalTooltip";
-import { useFetchTickets } from "@app/queries/tickets";
 
 dayjs.extend(utc);
 
@@ -78,7 +77,6 @@ export const MigrationWaves: React.FC = () => {
     useFetchMigrationWaves();
 
   const { trackers: trackers } = useFetchTrackers();
-  const { tickets: tickets } = useFetchTickets();
   const { data: applications } = useFetchApplications();
 
   const [migrationWaveModalState, setWaveModalState] = React.useState<
